Filter out deleted file mode lines from git-p4 diffs.
The recent fix for delete/rename support in git-p4 diffs still left a
`deleted file mode` line in the resulting diff. This was there before,
but is worth fixing now that the rest of delete handling works.

This is a trivial change to filter this line out, just like
`new file mode`.