Reduce log severity for VersionVault symlink warnings.
When processing directory changes involving symbolic links, we can end
up emitting a warning that we weren't able to find a file that looks
like "a -> b". This is expected and logging it is purely informational,
so we don't need it to be a warning. This change reduces that severity
to debug.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12664/