Fix incorrect check in custom client certificate loading.
The API implementation has the ability to load a custom client
certificate, but there was a small bug in this. The actual loading was
gated off by checking if either the cert or the key were available, but
we in fact need both. Changing an `or` to `and` fixes this.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/12671/