Fix rbt info on Python 3.
`rbt info` was depending on a `backports` module, which was only
installed on Python 2.7. This prevented the command from working at all
on Python 3.

This changes the usage to be a conditional import, using the modern
version if available and falling back on the backport.

Testing Done:
Successfully ran `rbt info` on Python 3.

Reviewed at https://reviews.reviewboard.org/r/12520/