Process ClearCase changeset items in the order listed.
When sanitizing an activity changeeset, we iterate through all changes,
build a dictionary, and then iterate through that. This can cause the
second iteration to be in a different order from the first on some
versions of Python. While probably harmless in normal use, this does
cause sporadic issues with test runs.

This change switches the dictionary to be an `OrderedDict`, so that the
order is preserved.

Testing Done:
Unit tests pass on all supported versions of Python.

Reviewed at https://reviews.reviewboard.org/r/12717/