Fix creating diffs for UCM activities whose changes are in other VOBs.
One of our more recent changes to improve the versioning of elements in
ClearCase/VersionVault introduced a problem where if the element was not inside
the same VOB as the CWD, we'd hit an error trying to query for its version. This
change adds the VOB UUID as a part of the path that we query, allowing this to
work across VOBs.

Testing Done:
Posted a UCM activity where all of the changes existed in a VOB other than the
one in my CWD. Before this change, I was getting an error executing cleartool.
After, the diff is posted successfully.

Reviewed at https://reviews.reviewboard.org/r/12894/