Update workflow docs to summarize steps and improve consistency.
This gives each of the RBTools workflow docs an overview summary in the
opening section, helping give people a broad sense of the steps that
will be covered.

It fixes some consistency issues with the opening step, header casing,
and some language.

It also future-proofs the instructions for closing a review request for
Review Board 6.

Testing Done:
Built the docs. Checked that all affected links worked and that there
were no build errors.

Reviewed at https://reviews.reviewboard.org/r/12954/