Update the Windows installer for the new code sign cert and Python 3.10.
We've recently renewed our code signing cert for Windows. This change
updates the installer to provide the new fingerprint for the cert, so we
can perform builds again.

It also updates our version of Python 3.10 to the latest 3.10.11
release.

Testing Done:
Successfully built a new installer and tested it (along with the new bundled
diff.exe support) on a Windows deployment.

Reviewed at https://reviews.reviewboard.org/r/12978/