Fix a regression caused by an over-indent in the recent Git changes.
The recent change to move Git over to `run_process` resulted in a block
of code that was indented too far, causing a fallback codepath for Git
repository lookup to fail. This broke tests and could break looking up
repositories.

This is a trivial fix that unindents the code block.