Move all base command support into rbtools.commands.base.
Modules in `rbtools.commands.base` now provide all the base command
classes, option processing, output handling, and errors. This cleans up
the `rbtools.commands` module and gives us room for further expansion.

This work was planned for RBTools 4, but had to slip.

With the exception of the, well, exception classes, all classes use
`housekeeping.ClassMovedMixin` to point to the new class. We don't
bother subclassing the old exceptions because catching a subclass
wouldn't do any good. We can't easily deprecate those, for the
time-being, we'll leave them in, directly assigning from the new
modules.

The moved classes have not been modernized. Aside from imports and
version change information in the docstrings, no changes have been made
to any of the moved code.

Testing Done:
Unit tests pass, except for one that's sensitive to generated output
and is seeing a warning. This will be addressed in a follow-up change.

Reviewed at https://reviews.reviewboard.org/r/13021/