Update code to consume the new RBToolsConfig for config access.
`RBToolsConfig` is now used throughout the codebase.
`BaseCommand.config` is now an instance of this, options now take this,
and base command support accesses state on this without the need for
conditional checks.

Options benefit the most. We can now get rid of all the custom
`default=` arguments on any option backed by a config key. The option
support has always based the default on the config, but only to reflect
values specified in config files. Since `RBToolsConfig` now manages all
defaults, the ones on options can be removed. That will happen in
another change, but the classes are prepared for it.

Testing Done:
Unit tests pass.

I've been using this locally, and have used it to post this change and
other related changes today.

Reviewed at https://reviews.reviewboard.org/r/13037/