Add stamping and fix configuration for the Perforce trigger script.
The Perforce trigger script can now stamp changesets with the review
request URL when the change is approved and submitted. This will only be
stamped if not already stamped.

There are also fixes for configuring the API user/token used for
communicating with Review Board.

Testing Done:
Tested all combinations of options locally.

Verified the fixes and improvements with a customer.

Reviewed at https://reviews.reviewboard.org/r/13046/