Fix a unit test regression in rbtools.config on Python 3.7-3.9.
The check for a Syntax Error when loading `.reviewboardrc` failed on
Python 3.7-3.9. This was due to a difference in how these versions
reported the column information.

This change makes the expected column conditional based on Python
version, and simplifies the message prefix we check to simply leave out
that information.

Testing Done:
All unit tests pass on Python 3.7-3.11.

Reviewed at https://reviews.reviewboard.org/r/13073/