Add a unit test function for overriding the environment.
This introduces a `TestCase.env` context manager that takes in a
dictionary of new environment settings, sets them, and then unsets them
after the context is run.

Through this, we should be able to safely work with the environment
without tests having to take special care to reset state.

Testing Done:
Made use of this in upcoming unit tests.

Reviewed at https://reviews.reviewboard.org/r/13070/