Conditionally return total_results from list resources.
We have some resources that behave like lists but don't contain a
`total_results` field. When we access this, we crash, because we assume
this will always be present.

This change makes that field optional, setting it to `None` if not
found.

Testing Done:
Unit tests pass.

Tested this with the watched review request resource, which exhibits
this failure.

Reviewed at https://reviews.reviewboard.org/r/13274/