Display useful error messages when passing invalid/expired API tokens to commands.
This displays useful error messages when passing invalid or expired API tokens
to a command. Previously, if you passed an API token that was expired or
invalidated it would silently fail and then prompt the user to enter a
username and password.

Testing Done:
- Tested `rbt api-get`, and `rbt login` commands with
  the `--api-token` option. Used valid, invalid and expired tokens.
- Tested the above commands with valid and invalid `--username` and
  `--password` options.

Reviewed at https://reviews.reviewboard.org/r/13128/