Add get_file_content APIs for Perforce and improve revision range diffs.
This change makes two improvements to the Perforce client backend in
RBTools. First, this implements the new get_file_content and
get_file_size APIs for use with uploading binary files to diffs. Second,
this plumbs through the tip revision for each file to the diff method so
that we can include it in the file headers. This was important when
posting submitted changesets so that we would have revision information
for the modified file as well as the original.

Testing Done:
- Posted a variety of submitted and pending changesets, including binary
  files.
- Verified that the revision information for modified files was present
  in the diff headers when posting submitted changesets.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13593/