Remove deprecated find_server_repository_info methods.
We used to have a huge hack where RBTools would do a bunch of stuff to
try to find the local repository, but then the client could replace the
entire repository info structure. In 3.0 we cleaned all this up to add
the `update_from_remote` method, which would let the client fetch
whatever data it needed. When we did this we kept the old methods around
with a deprecation warning because third party tools and scripts could
potentially still be calling them. This change finally removes the
methods.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13595/