Remove the last of code marked for removal in 4.0.
This removes the last of the stuff slated for removal in RBTools 4.0
(which we didn't actually remove in 4.0). The only silly thing here was
the submit_as argument to guess_existing_review_request, which had been
marked as required for 4.0 but we weren't actually passing in values
from most callsites.

Testing Done:
- Ran unit tests.
- Did a bunch of stuff that required guessing the review request from
  `rbt land`, `rbt post`, and `rbt stamp`.

Reviewed at https://reviews.reviewboard.org/r/13599/